+++ index.js
@@ -72,6 +72,11 @@ function session(options){
     , storeReady = true
     , rollingSessions = options.rolling || false;
 
+  // TODO: switch default to false on next major
+  var resaveSession = options.resave === undefined
+    ? true
+    : false;
+
   // notify user that this store is not
   // meant for a production environment
   if ('production' == env && store instanceof MemoryStore) {
@@ -153,7 +158,7 @@ function session(options){
             return
           }
         // compare hashes and ids
-        } else if (originalHash == hash(req.session) && originalId == req.session.id) {
+        } else if (!isModified(req.session)) {
           debug('unmodified session');
           return
         }
@@ -170,18 +175,28 @@ function session(options){
     res.end = function(data, encoding){
       res.end = end;
       if (!req.session) return res.end(data, encoding);
-      debug('saving');
       req.session.resetMaxAge();
-      req.session.save(function(err){
-        if (err) console.error(err.stack);
-        debug('saved');
-        res.end(data, encoding);
-      });
+
+      if (resaveSession || isModified(req.session)) {
+        debug('saving');
+        return req.session.save(function(err){
+          if (err) console.error(err.stack);
+          debug('saved');
+          res.end(data, encoding);
+        });
+      }
+
+      res.end(data, encoding);
     };
 
     // generate the session
     function generate() {
       store.generate(req);
+    }
+
+    // check if session has been modified
+    function isModified(sess) {
+      return originalHash != hash(sess) || originalId != sess.id;
     }
 
     // get the sessionID from the cookie