+++ History.md
@@ -1,3 +1,8 @@
+unreleased
+==========
+
+  * Add `resave` option to control saving unmodified sessions
+
 1.1.0 / 2014-05-12
 ==================